backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
Django>=3.2
